+++ kernel/net/ipv4/raw.h
@@ -22,7 +22,7 @@ void raw4_accept_packet(int protocol, struct net_buffer* nbuffer);
 void raw_ip4_put_to_rx_queue(struct socket* sock, struct net_buffer* nbuffer);
 
 int sock_raw4_create (struct socket* sock);
-int sock_raw4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
+ssize_t sock_raw4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
 ssize_t sock_raw4_recvfrom(struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
 int sock_raw4_close(struct socket* sock);
 short sock_raw4_poll(struct socket *sock, struct file *file, struct poll_ctl *pctl);
